+++ R/light_performance.R
@@ -80,8 +80,7 @@ light_performance.flashlight <- function(x, data = x$data, by = x$by,
     )
   }
   data[[label_name]] <- x$label
-  data <- dplyr::group_by(data, dplyr::across(tidyselect::all_of(c(label_name, by))))
-  agg <- dplyr::summarize(data, core_fun(dplyr::cur_data()), .groups = "drop")
+  agg <- Reframe(data, FUN = core_fun, .by = c(label_name, by))
 
   # Organize output
   add_classes(list(data = agg, by = by), classes = c("light_performance", "light"))